Moved image resizer to oscar.apps.dynamic_images, this is so that django picks up the app correctly for running tests.
Started work on unit tests for resizer, started with the URL parsing and fixed a couple of bugs in the process. Just begun work on testing the various mods.
Changed format conversion syntax to /filename.jpg.to.gif so that there's a greater tolerance for source files with dots in their names