:scream: Quarantine the evil before it is killed
Found horrendous javascript that renders the star ratings in product
listings.  This will be refactored so the appropriate ratings are passed
into the templates and don't have to be calculated by rendered DOM
elements then immediately deleting them.

For now, the badness is moved into its own method to make its
destruction easier later on.  Issue #420