Split plugins.js into constituent parts
I think plugins.js was a case of premature minificication.  Better to
store the plugins separately and use a compressor library to handle
concatenation and minification.

I put the plugins in their own folder for consistency with the plugins
that require extra assets (CSS, images).  Not sure if that's right.