ProductForm: fix bug when hiding related_products
The recent changes to the ProductUpdateView introduced a bug causing it
to break when the related_products field is hidden from the ProductForm.

This commit checks if the field is present before setting the queryset for
the related_products.
I also refactored the queryset out into a seperate function on
ProductForm to make it overridable.