Convert non-ASCII char to ASCII for category slugs
This uses the unidecode package to convert non-ASCII words to ASCII
equivalents.  We use a new slugify function to wrap this logic as we
will want to use it elsewhere.

See this StackOverflow thread for context:
http://stackoverflow.com/questions/702337/how-to-make-django-slugify-work-properly-with-unicode-strings/

Unidecode becomes a new dependency for Oscar.
https://pypi.python.org/pypi/Unidecode

Hat-tip to @avsazonov for suggesting this fix.