Update docs/ structure
To set a base for upcoming docs changes, the docs directory structure
has been streamlined to follow a simplified version of Django's docs.

The old _draft folder has been removed. Both files from there and the
current docs that don't fit nicely into the new structure are in _old/
for the moment.