Docs: Refactor translation howto
The old howto included out-dated info on how to translate Oscar. That
has been replaced by a link to the Transifex project.
It was the only howto left under the Contributing headline, which really
shouldn't be there (as there's an entirely different section in the docs
for contributing). As it is more of a topic anyway, and will hopefully
soon include documentation on translation style etc., it was turned into
a topic.