Introduce new API method for strategies (for basket lines)
This allows availability to be correctly determined for complex basket
lines which use basket attributes to hold information vital for
availability decisions.  An example is bundle products that store bundle
product IDs in the basket line attributes.

A new template tag is also introduced, which has an inconsistent name
with the product version.  This will be fixed in a later commit.