Preserve sort order in dashboard product search form
When you do a search in the product list, sort ordering is lost.

Sort order in the product list is kept in the query string – GET
variables. Add the current query string to the search form so that the
sort order is not reset when searching.

Fixes #1181

Issue #1195 is related.