Bump version of MySQL-python
This fixes problems installing the egg, e.g. in
https://travis-ci.org/tangentlabs/django-oscar/builds/18443119

I'm unsure of what causes the regression as the version of the package
was pinned before, so presume that either Travis or dependencies caused
the regression.
Nonetheless, bumping it to the latest stable version changes the issue.
I don't think further investigation is necessary at this point, if no
other issues arise.