Allow explicitly setting slug field even if saving for first time
E.g. Modelname.objects.create(slug='123', name='hi') would otherwise not
work, as add was True (first time save) and AutoSlugField's logic would
just create a slug from the name. It now checks if the slug field is set
already, and if it is, just returns the set slug.

It is important to note that explicitly specifying a slug does skip the
uniqueness validation.