Add checkout pre-condition for ensuring payment details are collected
This is required for the preview view which can be loaded via a GET
request normally. However, if a payment details view requires forms to
be submitted, then we want to ensure that you can't skip this step by
simply entering the URL of the preview view.

This change allows a `check_payment_data_is_captured` method to ensure
all information is entered before loading the preview view.