Disentangle Django's test client from Oscar's custom webtest class
Before this change, it was conflating WebTest functionality with the
Django test client. This change removes the Django test client
functionality (namely, the login function which was being automatically
called).

A shed-load of tests are adjusted to work with this change.

This almost completes the work for #826 - only a couple more references
to the Django client are left in the test codebase.