Remove jScrollPane
This Javascript plugin was used to style browser scrollbars in a
cross-platform way. I belief it's better to remove it:

a) It doesn't do anything, at least on my Ubuntu Chrome nor Firefox.
b) I'm sure there's decent reasons why there isn't a cross-platform
way to style scrollbars. It's a liability when it comes to smartphones
and tables, because those scrollbars use a different UI.
c) It's easy to add back again if so desired.