Add num_approved_reviews property to product model
This allows a more elegant way of avoiding unnecessary SQL queries when
looking up the number of reviews in templates.

This was spotted while investigating (and fixing) #1299. This
highlighted that the add-reviews view was not passing the reviews_count
variable which had been introduced in 9d75db402. For some reason, this
failed silently in using the en-gb locale but died noisily with
TypeError for ru and others.