Fix fake ORM definition for hand-patched migration
When I did
e5b3c5d6663fd5d5a330d22ccef70ca42fef5810, I changed my mind and, after
doing the bulk of the work, decided to not change the definitions for
ProductAttributeValues. At that point it seemed quicker to carefully
hand-edit the migrations to exclude the changes. I wasn't careful
enough.

This caused South to see a discrepance between the current model
definitions and the state of migrations.