Improve AddToBasketForm
Wowsers! This form must be seriously old - it still refers to products
as "item"s which was dropped from Oscar at about week 3. So it's overdue
a refactoring.

This change alters the constructor signature to not require the request
as that's a nasty codesmell as forms shouldn't know about the request
directly. The request is replaced by passing the basket - the strategy
instance can be accessed of the basket anyhow.

I've renamed the kwarg "instance" to "product" and fixed the callers.

I've also reordered methods in a few classes to roughly follow the
normal order of execution. I find this helps to understand classes you
haven't seen for a looooong time.