Restructure add-to-basket form and view
This is quite a large change. It:

- Alters the add-to-basket URL to have the "base" product's PK in it
- Changes the constructor of the AddToBasket form to not require
  purchase_info (even though I only just introduced it!)
- Renames the product_id field to variant_id and only uses it for group
  products

These changes are motivated by the variant workflow. Before this change,
you effectively had different forms being used for rending and
validation as the product being passed to the form constructor was the
group product when rendering but the variant when validating. This just
about worked but was not nice.

The purchase_info param is removed as we can only calculate the
purchase_info once the variant product has been validated. Hence we
again violate Demeter and make a deep call to the purchase info
calculation within the basket form. Might revisit this again.

This change also simplifies the basket view logic and allows a form to
be removed.