Allow products with a price of zero
Oscar now supports setting a price of zero for products.

Please note that currently a price of zero has no other special-casing
apart in a template. Notably, setting a currency is still required and
you can't add products to a basket of a different currency. This can
be improved if needed.

In the templates, the product price is displayed as "Free" instead
of zero (and whatever currency).

Writing an (initially failing) test exposed slightly incorrect handling
of the passed price in the stockrecord factory, as a price of zero
would always be replaced by the default price.

Fixes #1347.