Explicitly distinguish between standalone/parent/child products
Previously, the product "structure" has been determined implicitly. That
means that the decision if a product is a parent product or (what used
to be called) a variant product was made by looking at it's parent
field. This makes it harder to reason about it in edge cases, especially
when creating a parent product before creating child products.

This commit introduces a field to be able to explicitly set the intended
product structure. It also starts down the path of consistently naming
products either stand-alone, parent or child product.

Previous names for those products existed, but they have slight issues:
* "Canonical" will be an unknown word too many non-native English
  speakers. It also reminds of the makers of Ubuntu.
* "Group product" does not express that there's a hierarchy
  between parent and child products, and it's not clear to which level
  a group product belongs.
* "variant product" in itself is fine, but expressing the hierarchy by
  using "parent" and "child" is preferrable.