Repurpose Category.slug to store single slug instead of all slugs
Historically, the slug field contained both the category's slug, and the
slugs of it's ancestor's concatenated together. This leads to various
issues as detailed in #1670.

This commit greatly simplifies the logic by
* introducing a full_slug property, analogous to full_name, that
  returns what used to be stored in the slug field
* removing the complex 'update subtree' logic as it's not necessary any
  more
* Updating the ProductCategoryView's category fetching logic as needed

It feels like the data migration is scary, but I can't think of many
ways this could go wrong. We do need to be clear to instruct the users
that they must run it as well.
I haven't included a South data migration because we're scheduled to
remove Django 1.6 support per our policy.