Simplify category tag tests
Now that the category tag is greatly simplified, it's also easier to
test. The tests previously needed to render the template tag; now we can
just test the method itself. Together with some set operations, this
makes for simpler-to-understand and faster tests.

I also corrected a typo: 'Children' instead of 'Childrens'.